Move order helper into a helper file
