+++ imports/plugins/core/orders/client/containers/orderDashboardContainer.js
@@ -5,6 +5,7 @@ import { Tracker } from "meteor/tracker";
 import { Components, composeWithTracker } from "@reactioncommerce/reaction-components";
 import { Media, Orders, OrderSearch as OrderSearchCollection } from "/lib/collections";
 import { Reaction, i18next } from "/client/api";
+import { filterShippingStatus, filterWorkflowStatus } from "../helpers";
 import OrderDashboard from "../components/orderDashboard.js";
 import {
   PACKAGE_NAME,
@@ -14,94 +15,6 @@ import {
 } from "../../lib/constants";
 
 const shippingStrings = ["picked", "packed", "labeled", "shipped"];
-
-const OrderHelper =  {
-  makeQuery(filter) {
-    let query = {};
-
-    switch (filter) {
-      // New orders
-      case "new":
-        query = {
-          "workflow.status": "new"
-        };
-        break;
-
-      // Orders that have been approved
-      case "approved":
-        query = {
-          "workflow.status": "coreOrderWorkflow/processing",
-          "billing.paymentMethod.status": "approved"
-        };
-        break;
-
-      // Orders that have been captured
-      case "captured":
-        query = {
-          "billing.paymentMethod.status": "completed",
-          "shipping.shipped": false
-        };
-        break;
-
-      // Orders that are being processed
-      case "processing":
-        query = {
-          "workflow.status": "coreOrderWorkflow/processing"
-        };
-        break;
-
-      // Orders that are complete, including all items with complete status
-      case "completed":
-        query = {
-          "workflow.status": "coreOrderWorkflow/completed"
-        };
-        break;
-
-      case "canceled":
-        query = {
-          "workflow.status": "coreOrderWorkflow/canceled"
-        };
-        break;
-
-      default:
-    }
-
-    return query;
-  },
-  shippingFilter(filter) {
-    let query = {};
-
-    switch (filter) {
-      case "picked":
-        query = {
-          "shipping.workflow.status": "coreOrderWorkflow/picked"
-        };
-        break;
-
-      case "packed":
-        query = {
-          "shipping.workflow.status": "coreOrderWorkflow/packed"
-        };
-        break;
-
-      case "labelled":
-        query = {
-          "shipping.workflow.status": "coreOrderWorkflow/labeled"
-        };
-        break;
-
-      case "shipped":
-        query = {
-          "shipping.workflow.status": "coreOrderWorkflow/shipped"
-        };
-        break;
-
-      default:
-    }
-
-    return query;
-  }
-};
 
 class OrderDashboardContainer extends Component {
   static propTypes = {
@@ -203,7 +116,7 @@ class OrderDashboardContainer extends Component {
   }
 
   filterWorkflowStatus = (event, value) => {
-    const query = OrderHelper.makeQuery(value);
+    const query = filterWorkflowStatus(value);
 
     this.setState({
       query: { ...this.state.query, ...query },
@@ -212,7 +125,7 @@ class OrderDashboardContainer extends Component {
   }
 
   filterShippingStatus = (event, value) => {
-    const query = OrderHelper.shippingFilter(value);
+    const query = filterShippingStatus(value);
 
     this.setState({
       query: { ...this.state.query, ...query },
@@ -227,19 +140,19 @@ class OrderDashboardContainer extends Component {
 
     if (filterString === "workflow") {
       workflowFilter = "";
-      query = { ...OrderHelper.makeQuery(workflowFilter), ...OrderHelper.shippingFilter(shippingFilter.toLowerCase()) };
+      query = { ...filterWorkflowStatus(workflowFilter), ...filterShippingStatus(shippingFilter.toLowerCase()) };
 
       if (this.state.query.createdAt) {
         query.createdAt = this.state.query.createdAt;
       }
     } else if (filterString === "date") {
       query = {
-        ...OrderHelper.makeQuery(workflowFilter.toLowerCase()),
-        ...OrderHelper.shippingFilter(shippingFilter.toLowerCase())
+        ...filterWorkflowStatus(workflowFilter.toLowerCase()),
+        ...filterShippingStatus(shippingFilter.toLowerCase())
       };
     } else if (filterString === "shipping") {
       shippingFilter = "";
-      query = { ...OrderHelper.makeQuery(workflowFilter.toLowerCase()), ...OrderHelper.shippingFilter(shippingFilter) };
+      query = { ...filterWorkflowStatus(workflowFilter.toLowerCase()), ...filterShippingStatus(shippingFilter) };
 
       if (this.state.query.createdAt) {
         query.createdAt = this.state.query.createdAt;
